changed float to float32

+++ pointcloud/data.py
@@ -106,7 +106,7 @@ class PtCloudList(ItemList):
         self.copy_new.extend(['features', 'open_func', 'open_kwargs'])
         # TODO: ImageList sets 'self.c' here. Why?
 
-    def open(self, fn, ptcloud_class=PtCloudItem, dtype='float'):
+    def open(self, fn, ptcloud_class=PtCloudItem, dtype='float32'):
         item = self.open_func(fn, **self.open_kwargs)
 
         if isinstance(item, pd.DataFrame) or isinstance(item, pd.Series):
@@ -198,7 +198,7 @@ class PtCloudUpsampleLabelList(PtCloudList):
         super().__init__(items, **kwargs)
         self.loss_func = ChamferDistance()
 
-    def open(self, fn, ptcloud_class=PtCloudUpsampledItem, dtype='float'):
+    def open(self, fn, ptcloud_class=PtCloudUpsampledItem, dtype='float32'):
         return super().open(fn, ptcloud_class=ptcloud_class, dtype=dtype)
 
     def analyze_pred(self, pred:Tensor):
